Reorganize azure storage manager packages.
- This is done so that users of StorageManager do not need to
  vendor Azure SDK.
