+++ azure/storagemanager/azure_test.go
@@ -1,4 +1,4 @@
-package azure
+package storagemanager
 
 import (
 	"reflect"
